backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic (using versions with pre-built wheels)
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# LangChain
langchain==0.1.0
langchain-community==0.0.10

# Vector Store
chromadb==0.4.18

# Document Processing
pypdf==3.17.0

# Embeddings - Updated for compatibility
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

